requirements.txt
jmespath
pyparsing
capstone
nose 
coverage
pycryptodome
binwalk 
